ota_build: manage errors during the OTA build process

+++ ota_build/ota_build.sh
@@ -3,11 +3,29 @@
 cd $ANDROID_BUILD_TOP
 
 m installclean -j8
+
+if [ $? != 0 ]; then
+    echo "installclean failed"
+    exit 1
+fi
+
 m -j8
 
+if [ $? != 0 ]; then
+    echo "make failed"
+    exit 1
+fi
+
 rm -rf $OUT/ota_custom
+
 mkdir $OUT/ota_custom
 cp -r $ANDROID_BUILD_TOP/vendor/rs/ota_build/$TARGET_PRODUCT/ota_sample/* $OUT/ota_custom/
+
+if [ $? != 0 ]; then
+    echo "Cannot copy sample packages to out directory"
+    exit 1
+fi
+
 cd $OUT
 
 files=(
@@ -27,18 +45,59 @@ for file in "${files[@]}"; do
 done
 
 m vnod -j8
+
+if [ $? != 0 ]; then
+    echo "vnod failed"
+    exit 1
+fi
+
 m snod -j8
+
+if [ $? != 0 ]; then
+    echo "snod failed"
+    exit 1
+fi
+
 m bootimage -j8
+
+if [ $? != 0 ]; then
+    echo "make bootimage failed"
+    exit 1
+fi
+
 simg2img system.img system.raw
+
+if [ $? != 0 ]; then
+    echo "simg2img system failed"
+    exit 1
+fi
+
 simg2img vendor.img vendor.raw
+
+if [ $? != 0 ]; then
+    echo "simg2img vendor failed"
+    exit 1
+fi
+
 mv system.raw $OUT/ota_custom/system.img
 mv vendor.raw $OUT/ota_custom/vendor.img
 cp boot.img $OUT/ota_custom/boot.img
 
 cd $OUT/ota_custom
 zip -r ${TARGET_PRODUCT}_ota.zip *
+
+if [ $? != 0 ]; then
+    echo "Cannot compress final zip"
+    exit 1
+fi
+
 cd $OUT
 java -jar -Djava.library.path="$ANDROID_BUILD_TOP/out/host/linux-x86/lib64" $ANDROID_BUILD_TOP/out/host/linux-x86/framework/signapk.jar -w $ANDROID_BUILD_TOP/vendor/rs/config/security/releasekey.x509.pem $ANDROID_BUILD_TOP/vendor/rs/config/security/releasekey.pk8 $OUT/ota_custom/${TARGET_PRODUCT}_ota.zip $OUT/${TARGET_PRODUCT}_ota.zip
+
+if [ $? != 0 ]; then
+    echo "Cannot sign final zip"
+    exit 1
+fi
 
 rm -rf $OUT/ota_custom/${TARGET_PRODUCT}_ota.zip
